Split WidgetsGrid into displacement solver, drag hook, and component
computeDisplacedMoves and resizePushDirection move to displacement.ts
(pure, React-free), the drag monitor to use-drag-snap-position.ts, and
WidgetsGrid itself into its own folder per the component convention.
All code moved verbatim.
